Optimize 2024 day 6
When checking for cycles, advance to the next obstruction instead of a single step, and cache how many steps there are until the next obstruction.

For my input, the cache can be used ~90% of the time, and the hit rate is ~98% when it can be used.

This reduces runtime from ~6ms to ~1ms.