Update for #[target_feature] 1.1
target_features_11 was stabilized in 1.86 and allows safe functions to be marked with #[target_feature], simplifying the multiversion macro by removing the set of tt rules used to parse function definitions to inject the unsafe keyword and block.

In 1.87 architecture intrinsics that don't use raw pointers were made safe, allowing the removal of unsafe blocks in the SIMD implementations from the non-trait functions (trait implementations can't be marked with #[target_feature]).